+++ client/src/bin/bot/tui.rs
@@ -47,7 +47,7 @@ impl Tui {
     #[expect(clippy::needless_pass_by_ref_mut)]
     pub fn process_command(cmd: &str, command_s: &mut async_channel::Sender<Cmd>) -> bool {
         let matches = Command::new("veloren-botclient")
-            .version(common::util::DISPLAY_VERSION_LONG.as_str())
+            .version(common::util::DISPLAY_VERSION.as_str())
             .author("The veloren devs <https://gitlab.com/veloren/veloren>")
             .about("The veloren bot client allows logging in as a horde of bots for load-testing")
             .no_binary_name(true)