+++ server-cli/src/cli.rs
@@ -97,7 +97,7 @@ pub enum MessageReturn {
 #[derive(Parser)]
 #[command(
     name = "Veloren server TUI",
-    version = common::util::DISPLAY_VERSION_LONG.as_str(),
+    version = common::util::DISPLAY_VERSION.as_str(),
     about = "The veloren server tui allows sending commands directly to the running server.",
     author = "The veloren devs <https://gitlab.com/veloren/veloren>",
 )]
@@ -129,7 +129,7 @@ pub enum ArgvCommand {
 #[derive(Parser)]
 #[command(
     name = "Veloren server CLI",
-    version = common::util::DISPLAY_VERSION_LONG.as_str(),
+    version = common::util::DISPLAY_VERSION.as_str(),
     about = "The veloren server cli provides an easy to use interface to start a veloren server.",
     author = "The veloren devs <https://gitlab.com/veloren/veloren>",
 )]