+++ voxygen/src/menu/main/mod.rs
@@ -569,18 +569,17 @@ pub(crate) fn get_client_msg_error(
     // and server version it is almost definitely due to this mismatch rather than
     // a true networking error.
     let net_error = |error: String, mismatched_server_info: Option<ServerInfo>| -> String {
-        if let Some(server_info) =
-            mismatched_server_info.filter(|info| info.git_hash != *common::util::GIT_HASH)
-        {
+        if let Some(server_info) = mismatched_server_info.filter(|info| {
+            info.git_hash != *common::util::GIT_HASH
+                || info.git_timestamp != *common::util::GIT_TIMESTAMP
+        }) {
             format!(
-                "{} {}: {} ({}) {}: {} ({})",
+                "{} {}: {} {}: {}",
                 localization.get_msg("main-login-network_wrong_version"),
                 localization.get_msg("main-login-client_version"),
-                &*common::util::GIT_HASH,
-                &*common::util::GIT_DATE,
+                &*common::util::DISPLAY_VERSION,
                 localization.get_msg("main-login-server_version"),
-                server_info.git_hash,
-                server_info.git_date,
+                common::util::make_display_version(server_info.git_hash, server_info.git_timestamp),
             )
         } else {
             format!(